requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.4
fonttools==4.43.1
idna==3.4
imageio==2.31.6
importlib-resources==6.1.0
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.7.3
networkx==3.1
numpy==1.23.5
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
pip==23.3
pycocotools-windows==2.0.0.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
setuptools==68.0.0
six==1.16.0
tifffile==2023.7.10
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
typing_extensions==4.8.0
urllib3==2.0.7
wheel==0.41.2
zipp==3.17.0
